up to date

+++ django/core/management/commands/makemigrations.py
@@ -1,17 +1,20 @@
-import sys
 import os
-import operator
+import sys
+from itertools import takewhile
 
 from django.apps import apps
 from django.core.management.base import BaseCommand, CommandError
 from django.db.migrations import Migration
-from django.db.migrations.loader import MigrationLoader
 from django.db.migrations.autodetector import MigrationAutodetector
-from django.db.migrations.questioner import MigrationQuestioner, InteractiveMigrationQuestioner
+from django.db.migrations.loader import MigrationLoader
+from django.db.migrations.questioner import (
+    InteractiveMigrationQuestioner, MigrationQuestioner,
+    NonInteractiveMigrationQuestioner,
+)
 from django.db.migrations.state import ProjectState
 from django.db.migrations.writer import MigrationWriter
 from django.utils.six import iteritems
-from django.utils.six.moves import reduce
+from django.utils.six.moves import zip
 
 
 class Command(BaseCommand):
@@ -30,6 +33,8 @@ class Command(BaseCommand):
             help='Tells Django to NOT prompt the user for input of any kind.')
         parser.add_argument('-n', '--name', action='store', dest='name', default=None,
             help="Use this name for migration file(s).")
+        parser.add_argument('-e', '--exit', action='store_true', dest='exit_code', default=False,
+            help='Exit with error code 1 if no changes needing migrations are found.')
 
     def handle(self, *app_labels, **options):
 
@@ -38,7 +43,8 @@ class Command(BaseCommand):
         self.dry_run = options.get('dry_run', False)
         self.merge = options.get('merge', False)
         self.empty = options.get('empty', False)
-        self.migration_name = options.get('name', None)
+        self.migration_name = options.get('name')
+        self.exit_code = options.get('exit_code', False)
 
         # Make sure the app they asked for exists
         app_labels = set(app_labels)
@@ -63,10 +69,10 @@ class Command(BaseCommand):
 
         # If app_labels is specified, filter out conflicting migrations for unspecified apps
         if app_labels:
-            conflicts = dict(
-                (app_label, conflict) for app_label, conflict in iteritems(conflicts)
+            conflicts = {
+                app_label: conflict for app_label, conflict in iteritems(conflicts)
                 if app_label in app_labels
-            )
+            }
 
         if conflicts and not self.merge:
             name_str = "; ".join(
@@ -88,11 +94,15 @@ class Command(BaseCommand):
         if self.merge and conflicts:
             return self.handle_merge(loader, conflicts)
 
+        if self.interactive:
+            questioner = InteractiveMigrationQuestioner(specified_apps=app_labels, dry_run=self.dry_run)
+        else:
+            questioner = NonInteractiveMigrationQuestioner(specified_apps=app_labels, dry_run=self.dry_run)
         # Set up autodetector
         autodetector = MigrationAutodetector(
             loader.project_state(),
             ProjectState.from_apps(apps),
-            InteractiveMigrationQuestioner(specified_apps=app_labels, dry_run=self.dry_run),
+            questioner,
         )
 
         # If they want to make an empty migration, make one for each app
@@ -100,10 +110,10 @@ class Command(BaseCommand):
             if not app_labels:
                 raise CommandError("You must supply at least one app label when using --empty.")
             # Make a fake changes() result we can pass to arrange_for_graph
-            changes = dict(
-                (app, [Migration("custom", app)])
+            changes = {
+                app: [Migration("custom", app)]
                 for app in app_labels
-            )
+            }
             changes = autodetector.arrange_for_graph(
                 changes=changes,
                 graph=loader.graph,
@@ -120,15 +130,20 @@ class Command(BaseCommand):
             migration_name=self.migration_name,
         )
 
-        # No changes? Tell them.
-        if not changes and self.verbosity >= 1:
-            if len(app_labels) == 1:
-                self.stdout.write("No changes detected in app '%s'" % app_labels.pop())
-            elif len(app_labels) > 1:
-                self.stdout.write("No changes detected in apps '%s'" % ("', '".join(app_labels)))
+        if not changes:
+            # No changes? Tell them.
+            if self.verbosity >= 1:
+                if len(app_labels) == 1:
+                    self.stdout.write("No changes detected in app '%s'" % app_labels.pop())
+                elif len(app_labels) > 1:
+                    self.stdout.write("No changes detected in apps '%s'" % ("', '".join(app_labels)))
+                else:
+                    self.stdout.write("No changes detected")
+
+            if self.exit_code:
+                sys.exit(1)
             else:
-                self.stdout.write("No changes detected")
-            return
+                return
 
         self.write_migration_files(changes)
 
@@ -150,7 +165,7 @@ class Command(BaseCommand):
                 if not self.dry_run:
                     # Write the migrations file to the disk.
                     migrations_directory = os.path.dirname(writer.path)
-                    if not directory_created.get(app_label, False):
+                    if not directory_created.get(app_label):
                         if not os.path.isdir(migrations_directory):
                             os.mkdir(migrations_directory)
                         init_path = os.path.join(migrations_directory, "__init__.py")
@@ -187,24 +202,18 @@ class Command(BaseCommand):
                 migration = loader.get_migration(app_label, migration_name)
                 migration.ancestry = loader.graph.forwards_plan((app_label, migration_name))
                 merge_migrations.append(migration)
-            common_ancestor = None
-            for level in zip(*[m.ancestry for m in merge_migrations]):
-                if reduce(operator.eq, level):
-                    common_ancestor = level[0]
-                else:
-                    break
-            if common_ancestor is None:
+            all_items_equal = lambda seq: all(item == seq[0] for item in seq[1:])
+            merge_migrations_generations = zip(*[m.ancestry for m in merge_migrations])
+            common_ancestor_count = sum(1 for common_ancestor_generation
+                                        in takewhile(all_items_equal, merge_migrations_generations))
+            if not common_ancestor_count:
                 raise ValueError("Could not find common ancestor of %s" % migration_names)
             # Now work out the operations along each divergent branch
             for migration in merge_migrations:
-                migration.branch = migration.ancestry[
-                    (migration.ancestry.index(common_ancestor) + 1):
-                ]
-                migration.merged_operations = []
-                for node_app, node_name in migration.branch:
-                    migration.merged_operations.extend(
-                        loader.get_migration(node_app, node_name).operations
-                    )
+                migration.branch = migration.ancestry[common_ancestor_count:]
+                migrations_ops = (loader.get_migration(node_app, node_name).operations
+                                  for node_app, node_name in migration.branch)
+                migration.merged_operations = sum(migrations_ops, [])
             # In future, this could use some of the Optimizer code
             # (can_optimize_through) to automatically see if they're
             # mergeable. For now, we always just prompt the user.
@@ -222,7 +231,7 @@ class Command(BaseCommand):
                     for migration in merge_migrations
                 ]
                 try:
-                    biggest_number = max([x for x in numbers if x is not None])
+                    biggest_number = max(x for x in numbers if x is not None)
                 except ValueError:
                     biggest_number = 1
                 subclass = type("Migration", (Migration, ), {
@@ -230,7 +239,18 @@ class Command(BaseCommand):
                 })
                 new_migration = subclass("%04i_merge" % (biggest_number + 1), app_label)
                 writer = MigrationWriter(new_migration)
-                with open(writer.path, "wb") as fh:
-                    fh.write(writer.as_string())
-                if self.verbosity > 0:
-                    self.stdout.write("\nCreated new merge migration %s" % writer.path)
+
+                if not self.dry_run:
+                    # Write the merge migrations file to the disk
+                    with open(writer.path, "wb") as fh:
+                        fh.write(writer.as_string())
+                    if self.verbosity > 0:
+                        self.stdout.write("\nCreated new merge migration %s" % writer.path)
+                elif self.verbosity == 3:
+                    # Alternatively, makemigrations --merge --dry-run --verbosity 3
+                    # will output the merge migrations to stdout rather than saving
+                    # the file to the disk.
+                    self.stdout.write(self.style.MIGRATE_HEADING(
+                        "Full merge migrations file '%s':" % writer.filename) + "\n"
+                    )
+                    self.stdout.write("%s\n" % writer.as_string())
